+++ packages/workbox-background-sync/QueuePlugin.mjs
@@ -13,7 +13,7 @@
  limitations under the License.
 */
 
-import '../_version.mjs';
+import './_version.mjs';
 
 /**
  * A class implementing the `fetchDidFail` lifecycle callback. This makes it
