+++ packages/workbox-background-sync/models/QueueStore.mjs
@@ -14,7 +14,7 @@
 */
 
 import StorableRequest from './StorableRequest.mjs';
-import {DB_NAME, OBJECT_STORE_NAME, INDEXED_PROP} from './constants.mjs';
+import {DB_NAME, OBJECT_STORE_NAME, INDEXED_PROP} from '../utils/constants.mjs';
 import '../_version.mjs';
 
 /**
