+++ packages/workbox-background-sync/utils/constants.mjs
