+++ packages/workbox-runtime-caching/StaleWhileRevalidate.mjs
@@ -13,7 +13,9 @@
  limitations under the License.
 */
 
-import {_private} from 'workbox-core';
+import {getRuntimeName} from 'workbox-core/_private/cacheNames.mjs';
+import cacheWrapper from 'workbox-core/_private/cacheWrapper.mjs';
+import fetchWrapper from 'workbox-core/_private/fetchWrapper.mjs';
 
 import cacheOkAndOpaquePlugin from './plugins/cacheOkAndOpaquePlugin.mjs';
 import './_version.mjs';
@@ -43,8 +45,7 @@ class StaleWhileRevalidate {
    * conjunction with this caching strategy.
    */
   constructor(options = {}) {
-    this._cacheName =
-      _private.cacheNames.getRuntimeName(options.cacheName);
+    this._cacheName = getRuntimeName(options.cacheName);
       this._plugins = options.plugins || [];
 
     if (options.plugins) {
@@ -76,13 +77,13 @@ class StaleWhileRevalidate {
       // core.assert.isInstance({event}, FetchEvent);
     }
 
-    const fetchAndCachePromise = _private.fetchWrapper.fetch(
+    const fetchAndCachePromise = fetchWrapper.fetch(
       event.request,
       null,
       this._plugins
     )
     .then(async (response) => {
-      await _private.cacheWrapper.put(
+      await cacheWrapper.put(
         this._cacheName,
         event.request,
         response.clone(),
@@ -92,7 +93,7 @@ class StaleWhileRevalidate {
       return response;
     });
 
-    const cachedResponse = await _private.cacheWrapper.match(
+    const cachedResponse = await cacheWrapper.match(
       this._cacheName,
       event.request,
       null,
