+++ packages/workbox-runtime-caching/_public.mjs
@@ -14,13 +14,16 @@
   limitations under the License.
 */
 
-/**
- * There are common caching strategies that most service workers will need
- * and use. This module provides simple implementations of these strategies.
- *
- * @module workbox-runtime-caching
- */
+import CacheFirst from './CacheFirst.mjs';
+import CacheOnly from './CacheOnly.mjs';
+import NetworkFirst from './NetworkFirst.mjs';
+import NetworkOnly from './NetworkOnly.mjs';
+import StaleWhileRevalidate from './StaleWhileRevalidate.mjs';
 
-import './_version.mjs';
-
-export * from './_public.mjs';
+export {
+  CacheFirst,
+  CacheOnly,
+  NetworkFirst,
+  NetworkOnly,
+  StaleWhileRevalidate,
+};
